requirements.txt
pycryptodomex==3.21.0
pytest==8.2.2
pytest-cov==5.0.0
